Phase 3: Communication hardening — checksum, handshake, version, int64 encoder
Firmware changes:
- Add XOR checksum (*XX suffix) to all outbound telemetry lines
- Validate inbound checksum when present, accept bare lines for bench compat
- Add NAVBOT_PARSE_BAD_CHECKSUM parse result
- PING now responds with firmware version: ACK PING 1.1.0
- Encoder count widened from int32 to int64 (eliminates 50-hour rollover)
- NAVBOT_PROTOCOL_MAX_LINE increased from 96 to 128 for checksum headroom

Pi serial bridge changes:
- Extract checksum utilities to standalone checksum.py (no ROS dependency)
- Append checksum to all outbound commands
- Validate checksum on all inbound telemetry, log and discard corrupt lines
- Reconnect handshake: flush buffer, send STOP+PING, verify ACK PING
- Track firmware version from PING response, log on connect
- Track checksum failure count

Test additions (27 new tests, 77 total):
- test_checksum.py: XOR computation, roundtrip, corruption detection,
  known test vectors matching C algorithm
- test_serial_parser.py: checksum integration tests (valid, wrong,
  truncated, corrupt, no-checksum backward compat)

+++ firmware/makerpi_rp2040_base/include/wheel.h
@@ -21,8 +21,8 @@ typedef struct wheel {
     uint pin_rev;
     bool swap_dir;
 
-    volatile int32_t enc_count;
-    int32_t prev_count;
+    volatile int64_t enc_count;
+    int64_t prev_count;
     int32_t counts_per_rev;
     float wheel_radius_m;
